fix(profi): implement 9-bit GGGRRRBBB palette and #FE bit 7 (GX0)
Widens profiPalette to a 9-bit GGGRRRBBB entry (Karabas video.vhd
205-208/219): the extra blue LSB is latched from #FE.D7 on the OUT
that supplies the palette index, and #FE reads now report GX0
(palette[idx](6) xor palette[idx](0)) in DS80, else 1. Reset palette
levels updated to match the Karabas defaults (4/7 non-bright, 6/7
bright) uniformly across all three channels.

+++ core/src/emulator/platform.h
@@ -1009,7 +1009,9 @@ struct EmulatorState
 	} tape;
 	
 	uint8_t comp_pal[0x10];
-	uint8_t profiPalette[0x10];			// Profi hi-res palette: raw ~A15..A8 (GGGRRRBB), index = {bright,G,R,B}
+	uint16_t profiPalette[0x10];			// Profi hi-res palette: 9-bit GGGRRRBBB (bits 8:6 G, 5:3 R, 2:0 B;
+											// B0 is the extra blue LSB latched from #FE.D7 on the previous OUT),
+											// index = {bright,G,R,B}
 	uint8_t ulaplus_cram[64];
 	uint8_t ulaplus_mode;
 	uint8_t ulaplus_reg;
